backend/auth-service/requirements.txt
flask
redis
pyjwt
psycopg2
